[micro] Move shrinking code into policy-specific file TestHeapSafety
